refactor(esl-utils): move `EventUtils.dispatch` implementation to separate `dispatchCustomEvent` method
