+++ src/modules/esl-event-listener/test/utils.test.ts
@@ -1,24 +1,6 @@
 import {EventUtils} from '../core';
 
 describe('dom/events: EventUtils', () => {
-  describe('dispatch', () => {
-    // TODO: more tests for dispatch method
-    test('dispatches event with custom event init on the provided element', () => {
-      const el = document.createElement('div');
-      jest.spyOn(el, 'dispatchEvent');
-
-      const eventName = `click${Math.random()}`;
-      const customEventInit = {detail: Math.random()};
-      EventUtils.dispatch(el, eventName, customEventInit);
-
-      expect(el.dispatchEvent).toHaveBeenCalled();
-
-      const event: CustomEvent = (el.dispatchEvent as jest.Mock).mock.calls[0][0];
-      expect(event.type).toBe(eventName);
-      expect((event as any).detail).toBe(customEventInit.detail);
-    });
-  });
-
   // TODO: extend tests + possibly move descriptor finding to separate module?
   describe('descriptors', () => {
     test('basic test 1', () => {
